refactor_built_in_models
