+++ iotdb-core/ainode/iotdb/ainode/core/model/sundial/configuration_sundial.py
@@ -63,3 +63,5 @@ class SundialConfig(PretrainedConfig):
         super().__init__(
             **kwargs,
         )
+
+# TODO: Lacking checkpoint_path
